pyproject.toml
[tool.poetry]
name = "basketball_reference_scraper"
version = "0.1.0"
description = ""
authors = ["Isak <isak.poirier@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.5.1"
bs4 = "^0.0.1"
numpy = "^1.23.4"
python-dateutil = "^2.8.2"
pytz = "^2022.5"
lxml = "^4.9.1"
requests = "^2.28.1"
six = "^1.16.0"
soupsieve = "^2.3.2"
Unidecode = "^1.3.6"
python-dotenv = "^0.21.0"


[tool.poetry.scripts]
data_scrape = "basketball_reference_scraper:start"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
